Got integration test working again after discovering a number of issues: cmd line parsing being wrong, version returned now includes hash, removed auth keys when open sourcing so it couldnt log in, quant now expects command field. Also broke it into multiple files for readability
